+++ Source/timeMan.hpp
@@ -16,7 +16,7 @@ namespace TimeMan{
 extern TimeType msecPerMove, msecInTC, nbMoveInTC, msecInc, msecUntilNextTC, overHead;
 extern TimeType targetTime, maxTime;
 extern DepthType moveToGo;
-extern unsigned long long maxNodes;
+extern uint64_t maxNodes;
 extern bool isDynamic;
 extern bool isUCIPondering;
 extern bool isUCIAnalysis;
